Fix typo

+++ src/meta/bcstm.c
@@ -25,7 +25,7 @@ VGMSTREAM * init_vgmstream_bcstm(STREAMFILE *streamFile) {
 	/* check header */
 	if ((uint32_t)read_32bitBE(0, streamFile) != 0x4353544D) /* "CSTM" */
 		goto fail;
-	if ((uint32_t)read_16bitLE(4, streamFile) != 0xFEFF)
+	if ((uint16_t)read_16bitLE(4, streamFile) != 0xFEFF)
 		goto fail;
 	
 	int section_count = read_16bitLE(0x10, streamFile);
